Mutilate Options.lua Phase 1
